+++ hpc/nvidia/cuda/fakeCUDA/cuda-module-mgmt.cpp
@@ -3,6 +3,9 @@
 #include <cuda.h>
 
 
+#ifdef __cplusplus
+extern "C" {
+#endif
 CUresult cuModuleLoad(CUmodule* module, const char *fname)
 {
 	return CUDA_SUCCESS;
@@ -12,3 +15,6 @@ CUresult cuModuleGetFunction(CUfunction* hfunc, CUmodule hmod, const char *name)
 {
 	return CUDA_SUCCESS;
 }
+#ifdef __cplusplus
+}
+#endif