+++ hpc/nvidia/cuda/fakeCUDA/types.h
@@ -8,5 +8,7 @@ typedef enum cudaMemcpyKind cudaMemcpyKind;
 typedef enum cudaMemoryAdvise cudaMemoryAdvise;
 typedef struct cudaMemLocation cudaMemLocation;
 
+#ifndef __cplusplus
 /* TODO: _Float32 may not supported */
 typedef _Float32 __half;
+#endif